requirements.txt
discord.py==2.6.4
flask==3.1.2
discord.py
flask
